requirements.txt
# PyTorch (install based on your CUDA version)
# CPU: pip install torch torchvision
# CUDA 11.8: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# CUDA 12.1: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
# Apple Silicon (MPS): pip install torch torchvision
torch>=2.0.0
torchvision>=0.15.0

# Model architecture
timm>=0.9.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bar
tqdm>=4.65.0

# Optional: Grad-CAM visualization
# opencv-python>=4.8.0
